+++ Source/Classes/Categories/NSData+ImageDetectors.h
@@ -11,6 +11,7 @@
 @interface NSData (PINImageDetectors)
 
 - (BOOL)pin_isGIF;
+- (BOOL)pin_isAnimatedGIF;
 #if PIN_WEBP
 - (BOOL)pin_isWebP;
 - (BOOL)pin_isAnimatedWebP;